move static APIProvider methods into helpers module
